Add Pipfile + Config For Faster Workflow
Pipfile generated from Pipenv to track dependencies. Furthermore,
the `config.py` file will allow for credentials to be read in,
and a connection to be established, automatically. The
reasoning for this being that instead of having to pass
through your credentials everytime it will happen by itself.

The only *gotcha* is that within the root directory a
`config.yaml` file will have to exist with the following params:

dbname: 'dbname'
host: 'host'
port: port
user: 'user'
password: 'password'